Implemented interaction between actuators and virtual farming environments

+++ src/main/java/no/ntnu/fieldnode/device/sensor/SDUSensorData.java
@@ -4,7 +4,7 @@ package no.ntnu.fieldnode.device.sensor;
  * Sensor data capturing a single Double-value of data with a given SI-unit.
  */
 public class SDUSensorData extends SensorData {
-    private final int value;
+    private final double value;
 
     private final String unit;
 
@@ -15,7 +15,7 @@ public class SDUSensorData extends SensorData {
      * @param value the value of the data
      * @param unit the unit of the data
      */
-    public SDUSensorData(int sensorAddress, int value, String unit) {
+    public SDUSensorData(int sensorAddress, double value, String unit) {
         super(sensorAddress);
         this.value = value;
         this.unit = unit;
